+++ frontend/controllers/portfolios/positions_controller.js
@@ -1,0 +1,9 @@
+import ApplicationController from "../application_controller"
+
+export default class extends ApplicationController {
+  filterPortfolios (event) {
+    event.preventDefault()
+
+    this.stimulate('PositionsReflex#index', document.getElementById('portfolio_id').value, document.getElementById('current_locale').value)
+  }
+}